Refactor codebase and update documentation
Refactored indentation and formatting in service and utility classes for AWS and GCS, as well as the backup pipeline, to improve readability and maintainability. Updated .gitignore to clean up and reorganize ignored files. Enhanced README.md and estrutura.txt for clearer project structure and instructions. Added 'pytest' and 'requests' to requirements.txt for testing and HTTP requests.